chapter-16/ai_agent_hardening.py
# ai_agent_hardening.py - Comprehensive AI Agent Security Hardening Framework
import os
import json
import yaml
import subprocess
import requests
import hashlib
import time
from datetime import datetime, timedelta
from typing import Dict, List, Optional, Tuple, Any
from dataclasses import dataclass, asdict
from enum import Enum
import sqlite3

class HardeningLevel(Enum):
    BASIC = "basic"
    ENHANCED = "enhanced"
    MAXIMUM = "maximum"
    PARANOID = "paranoid"

class ControlCategory(Enum):
    AUTHENTICATION = "authentication"
    AUTHORIZATION = "authorization"
    INPUT_VALIDATION = "input_validation"
    OUTPUT_FILTERING = "output_filtering"
    MONITORING = "monitoring"
    NETWORK_SECURITY = "network_security"
    DATA_PROTECTION = "data_protection"
    MODEL_INTEGRITY = "model_integrity"

@dataclass
class HardeningControl:
    control_id: str
    category: ControlCategory
    title: str
    description: str
    level: HardeningLevel
    automated: bool
    script_function: str
    validation_function: str
    compliance_frameworks: List[str]
    risk_reduction: int  # 1-10 scale

class AIAgentHardeningFramework:
    def __init__(self, agent_config_path: str, output_dir: str = "./hardening_output"):
        self.agent_config_path = agent_config_path
        self.output_dir = output_dir
        self.controls = self.load_hardening_controls()
        self.results = []
        os.makedirs(output_dir, exist_ok=True)
    def load_hardening_controls(self) -> List[HardeningControl]:
        return [
            # AUTHENTICATION CONTROLS
            HardeningControl(
                control_id="AUTH_001",
                category=ControlCategory.AUTHENTICATION,
                title="Multi-Factor Authentication for Agent Access",
                description="Require MFA for all agent management interfaces and API access",
                level=HardeningLevel.BASIC,
                automated=True,
                script_function="implement_mfa_requirements",
                validation_function="validate_mfa_enforcement",
                compliance_frameworks=["NIST", "SOC2", "ISO27001"],
                risk_reduction=8
            ),
            HardeningControl(
                control_id="AUTH_002", 
                category=ControlCategory.AUTHENTICATION,
                title="Agent-to-Agent Authentication",
                description="Implement mutual authentication between all AI agents",
                level=HardeningLevel.ENHANCED,
                automated=True,
                script_function="setup_mutual_tls",
                validation_function="validate_agent_auth",
                compliance_frameworks=["NIST", "Zero Trust"],
                risk_reduction=9
            ),
            # AUTHORIZATION CONTROLS
            HardeningControl(
                control_id="AUTHZ_001",
                category=ControlCategory.AUTHORIZATION,
                title="Principle of Least Privilege for Agent Tools",
                description="Restrict agent tool access to minimum required permissions",
                level=HardeningLevel.BASIC,
                automated=True,
                script_function="implement_least_privilege",
                validation_function="validate_tool_permissions",
                compliance_frameworks=["NIST", "SOC2"],
                risk_reduction=7
            ),
            HardeningControl(
                control_id="AUTHZ_002",
                category=ControlCategory.AUTHORIZATION,
                title="Dynamic Permission Escalation Controls",
                description="Require human approval for agent privilege escalation",
                level=HardeningLevel.ENHANCED,
                automated=True,
                script_function="setup_escalation_controls",
                validation_function="validate_escalation_workflow",
                compliance_frameworks=["NIST", "SOX"],
                risk_reduction=8
            ),
            HardeningControl(
                control_id="INPUT_001",
                category=ControlCategory.INPUT_VALIDATION,
                title="Prompt Injection Protection",
                description="Implement multi-layer prompt injection detection and filtering",
                level=HardeningLevel.BASIC,
                automated=True,
                script_function="deploy_prompt_injection_filters",
                validation_function="test_injection_resistance",
                compliance_frameworks=["OWASP", "NIST"],
                risk_reduction=9
            ),
            HardeningControl(
                control_id="INPUT_002",
                category=ControlCategory.INPUT_VALIDATION,
                title="Adversarial Input Detection",
                description="Deploy ML-based adversarial input detection systems",
                level=HardeningLevel.ENHANCED,
                automated=True,
                script_function="setup_adversarial_detection",
                validation_function="validate_adversarial_protection",
                compliance_frameworks=["NIST"],
                risk_reduction=7
            ),
            # OUTPUT FILTERING CONTROLS
            HardeningControl(
                control_id="OUTPUT_001",
                category=ControlCategory.OUTPUT_FILTERING,
                title="Data Loss Prevention for Agent Outputs",
                description="Scan all agent outputs for sensitive data leakage",
                level=HardeningLevel.BASIC,
                automated=True,
                script_function="implement_output_dlp",
                validation_function="test_dlp_effectiveness",
                compliance_frameworks=["GDPR", "HIPAA", "SOX"],
                risk_reduction=8
            ),
            HardeningControl(
                control_id="OUTPUT_002",
                category=ControlCategory.OUTPUT_FILTERING,
                title="Response Validation and Sanitization",
                description="Validate and sanitize all agent responses before delivery",
                level=HardeningLevel.ENHANCED,
                automated=True,
                script_function="setup_response_validation",
                validation_function="test_response_sanitization",
                compliance_frameworks=["OWASP", "NIST"],
                risk_reduction=6
            ),
            # MONITORING CONTROLS
            HardeningControl(
                control_id="MON_001",
                category=ControlCategory.MONITORING,
                title="Comprehensive Agent Activity Logging",
                description="Log all agent decisions, tool calls, and interactions",
                level=HardeningLevel.BASIC,
                automated=True,
                script_function="setup_comprehensive_logging",
                validation_function="validate_log_coverage",
                compliance_frameworks=["SOC2", "ISO27001", "NIST"],
                risk_reduction=7
            ),
            HardeningControl(
                control_id="MON_002",
                category=ControlCategory.MONITORING,
                title="Behavioral Anomaly Detection",
                description="Detect unusual agent behavior patterns and decision anomalies",
                level=HardeningLevel.ENHANCED,
                automated=True,
                script_function="deploy_behavioral_monitoring",
                validation_function="test_anomaly_detection",
                compliance_frameworks=["NIST"],
                risk_reduction=8
            ),
            # NETWORK SECURITY CONTROLS
            HardeningControl(
                control_id="NET_001",
                category=ControlCategory.NETWORK_SECURITY,
                title="Agent Network Segmentation",
                description="Isolate agent networks with micro-segmentation",
                level=HardeningLevel.ENHANCED,
                automated=True,
                script_function="implement_network_segmentation",
                validation_function="validate_network_isolation",
                compliance_frameworks=["NIST", "Zero Trust"],
                risk_reduction=8
            ),
            HardeningControl(
                control_id="NET_002",
                category=ControlCategory.NETWORK_SECURITY,
                title="Encrypted Agent Communications",
                description="Encrypt all inter-agent and agent-to-service communications",
                level=HardeningLevel.BASIC,
                automated=True,
                script_function="setup_encrypted_communications",
                validation_function="validate_encryption_enforcement",
                compliance_frameworks=["NIST", "SOC2"],
                risk_reduction=7
            ),
            # DATA PROTECTION CONTROLS
            HardeningControl(
                control_id="DATA_001",
                category=ControlCategory.DATA_PROTECTION,
                title="Agent Memory Encryption",
                description="Encrypt agent memory and context stores at rest and in transit",
                level=HardeningLevel.ENHANCED,
                automated=True,
                script_function="implement_memory_encryption",
                validation_function="validate_memory_protection",
                compliance_frameworks=["GDPR", "HIPAA", "NIST"],
                risk_reduction=8
            ),
            HardeningControl(
                control_id="DATA_002",
                category=ControlCategory.DATA_PROTECTION,
                title="Data Retention and Purging Policies",
                description="Implement automated data lifecycle management for agent data",
                level=HardeningLevel.BASIC,
                automated=True,
                script_function="setup_data_lifecycle",
                validation_function="validate_data_purging",
                compliance_frameworks=["GDPR", "CCPA", "SOX"],
                risk_reduction=6
            ),
            # MODEL INTEGRITY CONTROLS
            HardeningControl(
                control_id="MODEL_001",
                category=ControlCategory.MODEL_INTEGRITY,
                title="AI Model Signing and Verification",
                description="Cryptographically sign all AI models and verify signatures",
                level=HardeningLevel.ENHANCED,
                automated=True,
                script_function="implement_model_signing",
                validation_function="validate_model_signatures",
                compliance_frameworks=["NIST", "Supply Chain Security"],
                risk_reduction=9
            ),
            HardeningControl(
                control_id="MODEL_002",
                category=ControlCategory.MODEL_INTEGRITY,
                title="Model Drift Detection and Alerting",
                description="Monitor for unauthorized model changes and performance drift",
                level=HardeningLevel.ENHANCED,
                automated=True,
                script_function="setup_drift_monitoring",
                validation_function="test_drift_detection",
                compliance_frameworks=["NIST"],
                risk_reduction=8
            )
        ]
# Additional controls and automation scripts would be defined below.
